pyproject.toml
[project]
name = "remote-mcp-servers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "supabase>=2.18.1",
]
